show upload progress
This requires that the actions removeFile and setFileProgress operate
based on the backingFile to find the fileObject because otherwise the
submitFile closure will contain the outdated fileObject which will
prevent the other actions from finding the fileObject by referential
equality.

Perhaps instead after a dispatch() we can getState() to update the
fileObject?