set CSRF-Token cookie on each request
This is what Angular does and it seems to make more sense since we get a
new CSRF Token more frequently than if we embed the CSRF token in the
template.

Angular docs note that only the same origin will be able to read the
cookie, which is necessary to set the X-CSRF-TOKEN header on a request.