reject on multicall errors
node-xmlrpc properly returns an error when a call returns an XML-RPC
error, but if we make a multicall to rtorrent, the call result (which
may be an error) is nested deep within the response, so node-xmlrpc is
not aware of this. Check for this, and for debuggability include the
methodName and its params that yielded that error.