handle situations where rtorrent is unavailable
On the server, the catch-all error handler checks to see if the error
has to do with failure to communicate with the RPC socket. If so, it
sends HTTP 503 SERVICE_UNAVAILABLE.

On the client, any request that indirectly communicates with rtorrent
will handle the situation where a 503 may be received. If it is, action
SET_RTORRENT_AVAILABILITY is triggered to reflect the fact that rtorrent
is not available. Conversely, on every successful request the
availability is set to true.