add download attribute to download links
It seems like not having this or target='_blank' would break the react
app, presumably at the react-router level or something.

Using a target attribute works but apparently opens up the possibility
of vulnerabilities. The download attribute works on all modern HTML5
browsers which I expect to be used for this anyway.